+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-## 1.0.2
+## 1.1.0
+
+_Feat_
+
+- [Issue #83](https://github.com/XPEHO/xpeapp_admin/issues/83) Possibility to open campaign
+- [Issue #82](https://github.com/XPEHO/xpeapp_admin/issues/82) Possibility to close campaign
 
 _Fix_